rss_feeds/requirements.txt
python-dotenv
beautifulsoup4
lxml
gotrue==1.0.2
supabase==1.0.3
python-dateutil
pytz
DateTime

pandas
scikit-learn~=1.3.0
numpy~=1.24.2
nltk~=3.8.1
spacy~=3.7.2
# postgrest~=0.10.8
gensim~=4.3.2
transformers~=4.34.1
torch~=2.1.0
textblob~=0.17.1
textacy
keybert~=0.8.3
keyphrase-vectorizers~=0.0.11